+++ XMapper.Tests/InitialPropertyListTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Xunit;
@@ -7,10 +8,47 @@ namespace XMapper.Tests;
 
 public class InitialPropertyListTests
 {
+    public class Dummy1
+    {
+        public string XString { get; set; } = "";
+        public string? XNullableString { get; set; }
+        public int XInt { get; set; }
+        public int? XNullableInt { get; set; }
+        public DummyEnum XEnum { get; set; }
+        public DummyEnum? XNullableEnum { get; set; }
+
+
+        public Type? XType { get; set; }
+        public IEnumerable<int> XEnumerable { get; set; } = Enumerable.Empty<int>();
+    }
+
+    public class Dummy2
+    {
+        public string XString { get; set; } = "";
+        public string? XNullableString { get; set; }
+        public string? XNullableString2 { get; set; }
+        public int XInt { get; set; }
+        public int? XNullableInt { get; set; }
+        public DummyEnum XEnum { get; set; }
+        public DummyEnum? XNullableEnum { get; set; }
+    }
+
+    public class Dummy3
+    {
+        public string? XNullableString { get; set; }
+    }
+
+    public enum DummyEnum
+    {
+        None,
+        One,
+    }
+
     private static List<PropertyInfo> GetPrivatePropertyInfos(object mapper)
     {
         return (List<PropertyInfo>)mapper.GetType().GetRuntimeFields().First(x => x.Name == "_propertyInfos").GetValue(mapper)!;
     }
+
 
     [Fact]
     public void SourceProperties_Dummy1()
@@ -21,11 +59,13 @@ public class InitialPropertyListTests
         Assert.Equal(new[]
         {
             nameof(Dummy1.XEnum),
+            nameof(Dummy1.XEnumerable),
             nameof(Dummy1.XInt),
             nameof(Dummy1.XNullableEnum),
             nameof(Dummy1.XNullableInt),
             nameof(Dummy1.XNullableString),
             nameof(Dummy1.XString),
+            nameof(Dummy1.XType),
         }, propertyNames);
     }