+++ XMapper/XMapper.cs
@@ -5,10 +5,9 @@ namespace XMapper;
 
 /// <summary>
 /// Create a mapper that can be used for mappings from <typeparamref name="TSource"/> to <typeparamref name="TTarget"/>.
-/// <para>Choose a <see cref="PropertyList"/>  (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which <see cref="ValueType"/> and <see cref="string"/> properties do you want to map automatically by name? You'll probably choose the class that has fewer properties.</para>
-/// <para>Use <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/> to ignore properties from that list.</para>
-/// <para>For ignored properties, you can do custom mappings via <see cref="IncludeAction"/>.</para>
-/// <para>For including reference type properties in the mapping, also use <see cref="IncludeAction"/> with another <see cref="XMapper{TSource, TTarget}"/>.</para>
+/// <para>Choose a <see cref="PropertyList"/>  (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which properties do you want to map automatically by name? You'll probably choose the class that has fewer properties.</para>
+/// <para>Use <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/> to ignore properties from that list. You should ignore all properties that do not have a matching property name or type.</para>
+/// <para>For ignored properties, you can do custom mappings via <see cref="IncludeAction"/> (both value type or reference type). In case of reference types, you may need to nest another <see cref="XMapper{TSource, TTarget}"/> here.</para>
 /// <para>
 /// Note 1: You can use fluent notation for all of this.<br />
 /// Note 2: For better performance you may want to assign the mapper to a static field (so that you can re-use it). If you're also mapping reference type members, then also assign those mappers to static fields.
@@ -22,10 +21,9 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
 
     /// <summary>
     /// Create a mapper that can be used for mappings from <typeparamref name="TSource"/> to <typeparamref name="TTarget"/>.
-    /// <para>Choose an <paramref name="initialPropertyList"/> (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which <see cref="ValueType"/> and <see cref="string"/> properties do you want to map automatically by name? You'll probably choose the class that has fewer properties.</para>
-    /// <para>Use <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/> to ignore properties from that list.</para>
-    /// <para>For ignored properties, you can do custom mappings via <see cref="IncludeAction"/>.</para>
-    /// <para>For including reference type properties in the mapping, also use <see cref="IncludeAction"/> with another <see cref="XMapper{TSource, TTarget}"/>.</para>
+    /// <para>Choose a <paramref name="initialPropertyList"/>  (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which properties do you want to map automatically by name? You'll probably choose the class that has fewer properties.</para>
+    /// <para>Use <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/> to ignore properties from that list. You should ignore all properties that do not have a matching property name or type.</para>
+    /// <para>For ignored properties, you can do custom mappings via <see cref="IncludeAction"/> (both value type or reference type). In case of reference types, you may need to nest another <see cref="XMapper{TSource, TTarget}"/> here.</para>
     /// <para>
     /// Note 1: You can use fluent notation for all of this.<br />
     /// Note 2: For better performance you may want to assign the mapper to a static field (so that you can re-use it). If you're also mapping reference type members, then also assign those mappers to static fields.
@@ -36,12 +34,11 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
         _propertyList = initialPropertyList;
         _propertyInfos = (_propertyList == PropertyList.Source ? typeof(TSource) : typeof(TTarget))
             .GetProperties()
-            .Where(x => x.PropertyType.IsValueType || x.PropertyType == typeof(string))
             .ToList();
     }
 
     /// <summary>
-    /// Only to use in case of <see cref="PropertyList.Source"/>. By default for all of <typeparamref name="TSource"/>'s <see cref="ValueType"/> and <see cref="string"/> properties an automatic mapping attempt is made (by equal name). If <typeparamref name="TTarget"/> does not have a matching property, you should ignore it here. You can also ignore a property and then include a custom mapping via <see cref="IncludeAction"/>.
+    /// Only to use in case of <see cref="PropertyList.Source"/>. By default for all of <typeparamref name="TSource"/>'s properties an automatic mapping attempt is made (by equal name). If <typeparamref name="TTarget"/> does not have a matching property (name and type), you should ignore it here. You can also ignore a property and then include a custom mapping via <see cref="IncludeAction"/>.
     /// </summary>
     public XMapper<TSource, TTarget> IgnoreSourceProperty<TProp>(Expression<Func<TSource, TProp>> propertySelector)
     {
@@ -54,7 +51,7 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
     }
 
     /// <summary>
-    /// Only to use in case of <see cref="PropertyList.Target"/>. By default for all of <typeparamref name="TTarget"/>'s <see cref="ValueType"/> and <see cref="string"/> properties an automatic mapping attempt is made (by equal name). If <typeparamref name="TSource"/> does not have a matching property, you should ignore it here. You can also ignore a property and then include a custom mapping via <see cref="IncludeAction"/>.
+    /// Only to use in case of <see cref="PropertyList.Target"/>. By default for all of <typeparamref name="TTarget"/>'s properties an automatic mapping attempt is made (by equal name). If <typeparamref name="TSource"/> does not have a matching property (name and type), you should ignore it here. You can also ignore a property and then include a custom mapping via <see cref="IncludeAction"/>.
     /// </summary>
     public XMapper<TSource, TTarget> IgnoreTargetProperty<TProp>(Expression<Func<TTarget, TProp>> propertySelector)
     {
@@ -68,8 +65,8 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
 
     /// <summary>
     /// For including custom mappings.
-    /// <para>For (<see cref="ValueType"/> or <see cref="string"/>) properties that you don't want to be mapped automatically you can define a custom mapping here. Ignore the automatic mapping attempt via <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/>.</para>
-    /// <para>For including reference type properties in the mapping, your can also use this method. For better performance, initialize the member mapper outside of this method (assuming your intention is to re-use mappers for better execution times).</para>
+    /// <para>For properties that you don't want to be mapped automatically you can define a custom mapping here. Ignore the automatic mapping attempt via <see cref="IgnoreSourceProperty"/> or <see cref="IgnoreTargetProperty"/>.</para>
+    /// <para>In case of reference type properties that have their own mapper: for better performance, initialize the member mapper outside of this method (assuming your intention is to re-use mappers for better execution times).</para>
     /// </summary>
     public XMapper<TSource, TTarget> IncludeAction(Action<TSource, TTarget> includeAction)
     {
@@ -119,12 +116,7 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
         foreach (var sourcePropertyInfo in _propertyInfos)
         {
             var targetPropertyInfo = targetType.GetProperty(sourcePropertyInfo.Name) ?? throw new Exception($"Property '{sourcePropertyInfo.Name}' was not found on {nameof(target)} '{typeof(TTarget).Name}'.");
-            var value = sourcePropertyInfo.GetValue(source);
-            if (value == null && targetPropertyInfo.PropertyType.IsValueType && Nullable.GetUnderlyingType(targetPropertyInfo.PropertyType) == null)
-            {
-                throw new Exception($"'{typeof(TSource).Name}.{sourcePropertyInfo.Name}' was null, but '{typeof(TTarget).Name}.{targetPropertyInfo.Name}' is not nullable.");
-            }
-            targetPropertyInfo.SetValue(target, value);
+            MapPropertyValue(source, target, sourcePropertyInfo, targetPropertyInfo);
         }
     }
 
@@ -134,33 +126,36 @@ public class XMapper<TSource, TTarget> where TSource : class, new() where TTarge
         foreach (var targetPropertyInfo in _propertyInfos)
         {
             var sourcePropertyInfo = sourceType.GetProperty(targetPropertyInfo.Name) ?? throw new Exception($"Property '{targetPropertyInfo.Name}' was not found on {nameof(source)} '{typeof(TSource).Name}'.");
-            var value = sourcePropertyInfo.GetValue(source);
-            if (value == null && targetPropertyInfo.PropertyType.IsValueType && Nullable.GetUnderlyingType(targetPropertyInfo.PropertyType) == null)
-            {
-                throw new Exception($"'{typeof(TSource).Name}.{sourcePropertyInfo.Name}' was null, but '{typeof(TTarget).Name}.{targetPropertyInfo.Name}' is not nullable.");
-            }
-            targetPropertyInfo.SetValue(target, value);
+            MapPropertyValue(source, target, sourcePropertyInfo, targetPropertyInfo);
         }
+    }
+
+    private static void MapPropertyValue(TSource source, TTarget target, PropertyInfo sourcePropertyInfo, PropertyInfo targetPropertyInfo)
+    {
+        var value = sourcePropertyInfo.GetValue(source);
+        if (value == null && (
+            targetPropertyInfo.PropertyType.IsValueType && Nullable.GetUnderlyingType(targetPropertyInfo.PropertyType) == null
+            || new NullabilityInfoContext().Create(targetPropertyInfo).WriteState != NullabilityState.Nullable))
+        {
+            throw new Exception($"'{typeof(TSource).Name}.{sourcePropertyInfo.Name}' was null, but '{typeof(TTarget).Name}.{targetPropertyInfo.Name}' is not nullable.");
+        }
+        targetPropertyInfo.SetValue(target, value);
     }
 }
 
 /// <summary>
-/// Choose an initial property list (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which <see cref="ValueType"/> and <see cref="string"/> properties do you want to map by name? You'll probably want to choose the class that has fewer properties.
+/// Choose an initial property list (<see cref="PropertyList.Source"/> or <see cref="PropertyList.Target"/>): which properties do you want to map by name? You'll probably want to choose the class that has fewer properties.
 /// </summary>
 public enum PropertyList
 {
     /// <summary>
     /// Take the property list of the source object and automatically map from these properties.<br/>
     /// Optionally use <see cref="XMapper{TSource, TTarget}.IgnoreSourceProperty"/> and <see cref="XMapper{TSource, TTarget}.IncludeAction"/> for custom mapping.<br/>
-    /// Note that only <see cref="ValueType"/> and <see cref="string"/> property types can be mapped automatically by name.<br/>
-    /// For including reference type properties, also use <see cref="XMapper{TSource, TTarget}.IncludeAction"/> but then for a call to <see cref="XMapper{TSource, TTarget}.Map(TSource, TTarget)"/>.
     /// </summary>
     Source,
     /// <summary>
     /// Take the property list of the target object and automatically map to these properties.<br/>
     /// Optionally use <see cref="XMapper{TSource, TTarget}.IgnoreTargetProperty"/> and <see cref="XMapper{TSource, TTarget}.IncludeAction"/> for custom mapping.<br/>
-    /// Note that only <see cref="ValueType"/> and <see cref="string"/> property types can be mapped automatically by name.<br/>
-    /// For including reference type properties, also use <see cref="XMapper{TSource, TTarget}.IncludeAction"/> but then for a call to <see cref="XMapper{TSource, TTarget}.Map(TSource, TTarget)"/>.
     /// </summary>
     Target
 }