+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [3.0.0] 2022-04-17
+
+### Breaking changes
+
+- All properties are mapped automatically by name, even reference types. So migrating from 2.0 to 3.0 may require adding a few calls to `IgnoreSourceProperty` or `IgnoreTargetProperty`. In case the target and source member have the same type, the reference is passed on.
+- Trying to map null to a not-nullable reference type results in an error. This a an extra safety check that is probably welcome.
+- .NET 6 is required instead of .NET 5, so that `NullabilityInfoContext` can be used for (nullable) reference types.
+
 ## [2.0.1] 2022-04-16
 
 - Improvements in the documentation that is available on hovering over the `XMapper` class or its methods.